--- packages/commercetools/nuxt/index.js
@@ -1,9 +1,0 @@
-import path from 'path';
-
-export default function (moduleOptions) {
-  this.options.router.middleware.push('commercetools');
-  this.addPlugin({
-    src: path.resolve(__dirname, './plugin.js'),
-    options: moduleOptions
-  });
-}
