+++ packages/commercetools/theme/nuxt.config.js
@@ -1,57 +1,5 @@
 import webpack from 'webpack';
 
-const integrationConfig = {
-  api: {
-    uri: 'https://api.commercetools.com/vsf-ct-dev/graphql',
-    authHost: 'https://auth.sphere.io',
-    projectKey: 'vsf-ct-dev',
-    clientId: 'xlea3xo3vcavMN5kmDlFP4nu',
-    clientSecret: 'JejrKtQgU_KkNxPn_96UEAaEoPocNFqy',
-    scopes: [
-      'create_anonymous_token:vsf-ct-dev',
-      'manage_my_orders:vsf-ct-dev',
-      'manage_my_profile:vsf-ct-dev',
-      'manage_my_shopping_lists:vsf-ct-dev',
-      'manage_my_payments:vsf-ct-dev',
-      'view_products:vsf-ct-dev',
-      'view_published_products:vsf-ct-dev'
-    ]
-  },
-  locale: 'en',
-  acceptLanguage: ['en', 'de'],
-  currency: 'USD',
-  country: 'US',
-  countries: [
-    { name: 'US',
-      label: 'United States' },
-    { name: 'AT',
-      label: 'Austria' },
-    { name: 'DE',
-      label: 'Germany' },
-    { name: 'NL',
-      label: 'Netherlands' }
-  ],
-  currencies: [
-    { name: 'EUR',
-      label: 'Euro' },
-    { name: 'USD',
-      label: 'Dollar' }
-  ],
-  locales: [
-    { name: 'en',
-      label: 'English' },
-    { name: 'de',
-      label: 'German' }
-  ],
-  cookies: {
-    currencyCookieName: 'vsf-currency',
-    countryCookieName: 'vsf-country',
-    localeCookieName: 'vsf-locale'
-  }
-};
-
-const localeNames = integrationConfig.locales.map(l => ({ code: l.name, file: `${l.name}.js`, iso: l.name }));
-
 export default {
   mode: 'universal',
   server: {
@@ -102,7 +50,55 @@ export default {
       composables: '@vue-storefront/commercetools'
     }],
     // @core-development-only-end
-    ['@vue-storefront/commercetools/nuxt', integrationConfig]
+    ['@vue-storefront/commercetools/nuxt', {
+      api: {
+        uri: 'https://api.commercetools.com/vsf-ct-dev/graphql',
+        authHost: 'https://auth.sphere.io',
+        projectKey: 'vsf-ct-dev',
+        clientId: 'xlea3xo3vcavMN5kmDlFP4nu',
+        clientSecret: 'JejrKtQgU_KkNxPn_96UEAaEoPocNFqy',
+        scopes: [
+          'create_anonymous_token:vsf-ct-dev',
+          'manage_my_orders:vsf-ct-dev',
+          'manage_my_profile:vsf-ct-dev',
+          'manage_my_shopping_lists:vsf-ct-dev',
+          'manage_my_payments:vsf-ct-dev',
+          'view_products:vsf-ct-dev',
+          'view_published_products:vsf-ct-dev'
+        ]
+      },
+      locale: 'en',
+      acceptLanguage: ['en', 'de'],
+      currency: 'USD',
+      country: 'US',
+      countries: [
+        { name: 'US',
+          label: 'United States' },
+        { name: 'AT',
+          label: 'Austria' },
+        { name: 'DE',
+          label: 'Germany' },
+        { name: 'NL',
+          label: 'Netherlands' }
+      ],
+      currencies: [
+        { name: 'EUR',
+          label: 'Euro' },
+        { name: 'USD',
+          label: 'Dollar' }
+      ],
+      locales: [
+        { name: 'en',
+          label: 'English' },
+        { name: 'de',
+          label: 'German' }
+      ],
+      cookies: {
+        currencyCookieName: 'vsf-currency',
+        countryCookieName: 'vsf-country',
+        localeCookieName: 'vsf-locale'
+      }
+    }]
   ],
   modules: [
     'nuxt-i18n',
@@ -122,19 +118,5 @@ export default {
         })
       })
     ]
-  },
-  i18n: {
-    locales: localeNames,
-    defaultLocale: localeNames[0].code,
-    lazy: true,
-    seo: true,
-    langDir: 'lang/',
-    vueI18n: {
-      fallbackLocale: localeNames[0].code
-    },
-    detectBrowserLanguage: {
-      cookieKey: integrationConfig.cookies.localeCookieName,
-      alwaysRedirect: true
-    }
   }
 };
