+++ packages/commercetools/theme/package.json
@@ -16,7 +16,6 @@
     "@storefront-ui/vue": "^0.7.10",
     "@vue-storefront/commercetools": "^0.0.3",
     "@vue-storefront/commercetools-api": "^0.0.3",
-    "@vue-storefront/commercetools-nuxt": "^0.0.1",
     "@vue-storefront/nuxt": "^0.0.3",
     "@vue-storefront/nuxt-theme": "^0.0.1",
     "cookie-universal-nuxt": "^2.1.3",
